:recycle: FlipCardCore로 분리
