+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flip_card/flip_card.dart';
-import 'package:flip_card_game/asset_name.dart';
+import 'package:flip_card_game/flip_card_core.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -31,37 +31,19 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  final _imageNames = [
-    AssetImageName.orange,
-    AssetImageName.banana,
-    AssetImageName.apple,
-    AssetImageName.strawberry,
-  ];
+  final flipCardCore = FlipCardCore();
 
-  final List<String> _randomImageNames = [];
-  final List<GlobalKey<FlipCardState>> _cardKeys = [];
-  int _frontCardCount = 0;
-  final List<int> _frontCardIndexes = [];
+  List<String> _randomImageNames = [];
+  List<GlobalKey<FlipCardState>> _cardKeys = [];
 
   @override
   void initState() {
     super.initState();
 
-    reset();
-  }
+    flipCardCore.reset();
 
-  void reset() {
-    // add 2 times
-    _randomImageNames.clear();
-    _randomImageNames.addAll(_imageNames);
-    _randomImageNames.addAll(_imageNames);
-
-    // shuffle
-    _randomImageNames.shuffle();
-
-    // create global key
-    _cardKeys.clear();
-    _cardKeys.addAll(_randomImageNames.map((_) => GlobalKey<FlipCardState>()));
+    _randomImageNames = flipCardCore.randomImageNames;
+    _cardKeys = flipCardCore.cardKeys;
   }
 
   @override
@@ -86,33 +68,8 @@ class _MyHomePageState extends State<MyHomePage> {
               }
               return FlipCard(
                 key: _cardKeys[index],
-                onFlip: () {
-                  _frontCardCount++;
-                  _frontCardIndexes.add(index);
-                },
-                onFlipDone: (bool) {
-                  if (_frontCardCount == 2) {
-                    _toggleCardToFront();
-
-                    _checkCardIsEqual();
-
-                    _frontCardCount = 0;
-
-                    if (_frontCardIndexes.length >= 2) {
-                      String firstCardName = _randomImageNames[_frontCardIndexes[0]];
-                      String secondCardName = _randomImageNames[_frontCardIndexes[1]];
-                      if (firstCardName == secondCardName) {
-                        _randomImageNames[_frontCardIndexes[0]] = '';
-                        _randomImageNames[_frontCardIndexes[1]] = '';
-
-                        setState(() {});
-                      }
-                    }
-
-                    _frontCardIndexes.clear();
-                    _frontCardCount = 0;
-                  }
-                },
+                onFlip: () => flipCardCore.onFlip(index),
+                onFlipDone: (bool) => flipCardCore.onFlipDone(setState),
                 front: Container(
                   width: 100,
                   height: 150,
@@ -134,38 +91,11 @@ class _MyHomePageState extends State<MyHomePage> {
       floatingActionButton: FloatingActionButton(
         onPressed: () {
           setState(() {
-            reset();
+            flipCardCore.reset();
           });
         },
         child: const Icon(Icons.refresh),
       ),
     );
-  }
-
-  void _toggleCardToFront() {
-    for (var cardKey in _cardKeys) {
-      if (cardKey.currentState == null) return;
-
-      if (!cardKey.currentState!.isFront) {
-        cardKey.currentState!.toggleCard();
-      }
-    }
-  }
-
-  void _checkCardIsEqual() {
-    print('_checkCardIsEqual');
-    print(_frontCardIndexes);
-    if (_frontCardIndexes.length >= 2) {
-      String firstCardName = _randomImageNames[_frontCardIndexes[0]];
-      String secondCardName = _randomImageNames[_frontCardIndexes[1]];
-      if (firstCardName == secondCardName) {
-        _randomImageNames[_frontCardIndexes[0]] = '';
-        _randomImageNames[_frontCardIndexes[1]] = '';
-
-        setState(() {});
-      }
-    }
-
-    _frontCardIndexes.clear();
   }
 }
